# microservices/iot/src/main/java/uk/ac/york/eng2/iot/dto/RoomDTO.java
package uk.ac.york.eng2.iot.dto;

import io.micronaut.serde.annotation.Serdeable;

@Serdeable
public class RoomDTO {
    private String name;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}