# microservices/reactive/src/main/openapi/iot-0.0.yml
openapi: 3.0.1
info:
  title: iot
  version: "0.0"
paths:
  /actuators:
    get:
      tags:
      - actuators
      operationId: list
      parameters:
      - name: type
        in: query
        explode: false
        schema:
          type: string
          nullable: true
      responses:
        "200":
          description: list 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actuator"
    post:
      tags:
      - actuators
      operationId: createActuator
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActuatorDTO"
        required: true
      responses:
        "200":
          description: createActuator 200 response
  /actuators/{id}:
    get:
      tags:
      - actuators
      operationId: getActuator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getActuator 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actuator"
    put:
      tags:
      - actuators
      operationId: updateActuator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActuatorDTO"
        required: true
      responses:
        "200":
          description: updateActuator 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actuator"
    delete:
      tags:
      - actuators
      operationId: deleteActuator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deleteActuator 200 response
  /actuators/{id}/state:
    put:
      tags:
      - actuators
      operationId: updateState
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActuatorStateDTO"
        required: true
      responses:
        "200":
          description: updateState 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actuator"
  /rooms:
    get:
      tags:
      - rooms
      operationId: list_1
      responses:
        "200":
          description: list_1 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      tags:
      - rooms
      operationId: createRoom
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomDTO"
        required: true
      responses:
        "200":
          description: createRoom 200 response
  /rooms/{id}:
    get:
      tags:
      - rooms
      operationId: getRoom
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getRoom 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    put:
      tags:
      - rooms
      operationId: updateRoom
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomDTO"
        required: true
      responses:
        "200":
          description: updateRoom 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      tags:
      - rooms
      operationId: deleteRoom
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deleteRoom 200 response
  /sensors:
    get:
      tags:
      - sensors
      operationId: list_2
      responses:
        "200":
          description: list_2 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sensor"
    post:
      tags:
      - sensors
      operationId: createSensor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorDTO"
        required: true
      responses:
        "200":
          description: createSensor 200 response
  /sensors/{id}:
    get:
      tags:
      - sensors
      operationId: getSensor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getSensor 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
    put:
      tags:
      - sensors
      operationId: updateSensor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorDTO"
        required: true
      responses:
        "200":
          description: updateSensor 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
    delete:
      tags:
      - sensors
      operationId: deleteSensor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deleteSensor 200 response
  /sensors/{id}/readings:
    get:
      tags:
      - sensors
      operationId: getReadings
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getReadings 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadingDTO"
components:
  schemas:
    Actuator:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        type:
          type: string
        room:
          $ref: "#/components/schemas/Room"
        targetState:
          type: string
    ActuatorDTO:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        roomId:
          type: integer
          format: int64
        targetState:
          type: string
    ActuatorStateDTO:
      type: object
      properties:
        targetState:
          type: string
    ReadingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sensorId:
          type: integer
          format: int64
        sensorName:
          type: string
        takenAt:
          type: string
          format: date-time
        value:
          type: number
          format: double
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        sensors:
          type: array
          items:
            $ref: "#/components/schemas/Sensor"
        actuators:
          type: array
          items:
            $ref: "#/components/schemas/Actuator"
    RoomDTO:
      type: object
      properties:
        name:
          type: string
    Sensor:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        type:
          type: string
        room:
          $ref: "#/components/schemas/Room"
    SensorDTO:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        roomId:
          type: integer
          format: int64